pyproject.toml
[project]
name = "visulinx-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.0",
    "boto3-stubs[s3]>=1.35.68",
    "boto3>=1.35.68",
    "fastapi[standard]>=0.115.5",
    "mypy>=1.13.0",
    "psycopg2-binary>=2.9.10",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.6.1",
    "pydantic[email]>=2.10.0",
    "pyjwt>=2.10.0",
    "python-multipart>=0.0.17",
    "sqlalchemy>=2.0.36",
    "testcontainers>=4.8.2",
    "python-magic>=0.4.27",
    "pytest-asyncio>=0.24.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.1",
    "freezegun>=1.5.1",
    "httpx>=0.27.2",
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "ruff>=0.7.4",
    "taskipy>=1.14.0",
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = "function"

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
dev = 'fastapi dev app/api.py'
migrate = 'alembic upgrade head'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
